CSS + PHP

+++ entroquizz/admin/create_theme.php
@@ -7,11 +7,11 @@
 ?>
 
 <section>
-		<h2>Cr&eacute;er un th&egrave;me</h2>
+		<h2>G&eacute;rer les th&egrave;me</h2>
 		
 		<article>
 			<h3>Ins&eacute;rer un th&egrave;me</h3>
-            <form class="creation" action="./create_theme.php" method="get">
+            <form class="creation-theme" action="./create_theme.php" method="get">
             	<fieldset>
             		<legend>Choisissez le th&egrave;me</legend>
             		
@@ -23,32 +23,37 @@
             </form>
             
             <?php 
-                $dbconn = pg_connect($confi);
                 
                 if(isset($_GET['theme'])){
                     $nom_theme = pg_escape_string($_GET['theme']);
-                    //Inserer les données dans la base de donnée
-                    $psql ="INSERT INTO Theme (nom_theme) VALUES ('".$nom_theme."')";
-                    //execution de la requete
-                    $req=pg_query($dbconn,$psql);
-    
-                    //Test si l'insertion a été effectué
-                   if (!$req) {
-                      echo pg_last_error();
-                   } else {
-                      echo "<p class=\"success\">Th&egrave;me : \"".$_GET['theme']."\" enregist&eacute; !</p>";
-                   }    
+                    
+                    if (is_in_database("Theme", "nom_theme", $nom_theme)) {
+                        echo "<p class=\"warning\">D&eacute;j&agrave; dans la base de donn&eacute;es</p>";
+                    } else {
+                        $dbconn = pg_connect($confi);
+                        //Inserer les données dans la base de donnée
+                        $psql ="INSERT INTO Theme (nom_theme) VALUES ('".$nom_theme."')";
+                        //execution de la requete
+                        $req=pg_query($dbconn,$psql);
+        
+                        //Test si l'insertion a été effectué
+                       if (!$req) {
+                          echo "<p class=\"error\">".pg_last_error()."</p>";
+                       } else {
+                          echo "<p class=\"success\">Th&egrave;me : \"".$_GET['theme']."\" enregist&eacute; !</p>";
+                       }    
+                        if(isset($_GET['themeN'])){
+                            $p = $_GET['themeN'];
+                            $req2 = " UPDATE Theme SET nom_theme = ".$p." WHERE id_theme = ".$_GET['updateid'];
+                            $req=pg_query($dbconn,$req2);
+                        }
+                        if (isset($_GET['delete'])) {
+                            $req1="DELETE FROM Theme WHERE id_theme = ".$_GET['delete'];
+                            $req=pg_query($dbconn,$req1);
+                        }
+                        pg_close($dbconn);
+                    }
                 }
-                if(isset($_GET['themeN'])){
-                    $p = $_GET['themeN'];
-                    $req2 = " UPDATE Theme SET nom_theme = ".$p." WHERE id_theme = ".$_GET['updateid'];
-                    $req=pg_query($dbconn,$req2);
-                }
-                if (isset($_GET['delete'])) {
-                    $req1="DELETE FROM Theme WHERE id_theme = ".$_GET['delete'];
-                    $req=pg_query($dbconn,$req1);
-                }
-                pg_close($dbconn);
             ?>
 		</article>
